Fix with EDA to only calculate the mean and std of non zero pixels + Improve the memory usage

+++ CNN/tester.py
@@ -13,11 +13,11 @@ def test(model, test_loader, criterion, device, output_norm_scaling=False, outpu
     for it, batch in enumerate(tqdm_iter):
         with torch.no_grad():
             X, pt, ieta, iphi, m = (
-                batch["X_jets"].float(),
-                batch["pt"].float(),
-                batch["ieta"].float(),
-                batch["iphi"].float(),
-                batch["m"].float(),
+                batch["X_jets"],
+                batch["pt"],
+                batch["ieta"],
+                batch["iphi"],
+                batch["m"],
             )
 
             X = X.to(device, non_blocking=True)
